// src/types.ts
export type Address =`0x${string}`;
export type Hex = `0x${string}`;
export type Hash = `0x${string}`;
export type EmptyHex = `0x`;
export type PromiseOrValue<T> = T | Promise<T>;
export type BigNumberish = string | bigint | number;
export type BytesLike = Uint8Array | string;


export interface SendUserOpResponse {
	userOpHash: Hash,
	transactionHash?: Hash 
}

export interface ClientOptions {
	chainId: number;
	privateKey?: string;
	rpcUrl?: string;
	isSponsoredTrx?: boolean
	noAccountCreateSponsorship?: boolean
	paymasterEndpoint?: string;
	paymaster?: "PIMLICO" | "STACKUP" | "MAGMAR";
	paymasterId?: string; //for magmar paymaster
	erc20token?: string;
	bundlerEndpoint?:string;
}

export interface TelegramClientOptions {
	chainId: number;
	telegramUserId: string;
	walletAddress: Address;
	rpcUrl?: string;
	isSponsoredTrx?: boolean
	paymasterEndpoint?: string;
	paymaster?: "PIMLICO" | "STACKUP",
	bundlerEndpoint?:string;
}

export interface UserOperationCallData {
	target: Address;
	data: Hex;
	value?: bigint;
}

export type BatchUserOperationCallData = UserOperationCallData[];
// ...
  export interface UserOperationReceiptObject {
	/* 32 Bytes - hash of the block where this log was in. null when its pending. null when its pending log */
	blockHash: Hash;
	/* The block number where this log was in. null when its pending. null when its pending log. */
	blockNumber: BigNumberish;
	/* The index of the transaction within the block. */
	transactionIndex: BigNumberish;
	/* 32 Bytes - hash of the transaction. null when its pending. */
	transactionHash: Hash;
	/* 20 Bytes - address of the sender */
	from: Address;
	/* 20 Bytes - address of the receiver. null when its a contract creation transaction */
	to: Address;
	/* The total amount of gas used when this transaction was executed in the block. */
	cumulativeGasUsed: BigNumberish;
	/* The amount of gas used by this specific transaction alone */
	gasUsed: BigNumberish;
	/* 20 Bytes - The contract address created, if the transaction was a contract creation, otherwise null */
	contractAddress: Address;
	logs: UserOperationReceiptLog[];
	/* 256 Bytes - Bloom filter for light clients to quickly retrieve related logs */
	logsBloom: Hex;
	/* 32 bytes of post-transaction stateroot. (pre Byzantium hard fork at block 4,370,000) */
	root: Hex;
	/* Either 1 (success) or 0 (failure). (post Byzantium hard fork at block 4,370,000) */
	status: number;
	/* The cumulative gas used in the block containing this UserOperation. */
	effectiveGasPrice: BigNumberish;
	/* The type of the recipt object */
	type: string;
  }
  

  export interface UserOperationReceiptLog {
	/* The hash of the block where the given transaction was included. */
	blockHash: Hash;
	/* The number of the block where the given transaction was included. */
	blockNumber: BigNumberish;
	/* The index of the transaction within the block. */
	transactionIndex: BigNumberish;
	/* 20 Bytes - address from which this log originated. */
	address: Address;
	/* Integer of the log index position in the block. null when its pending log. */
	logIndex: BigNumberish;
	/* Contains one or more 32 Bytes non-indexed arguments of the log. */
	data: Hex;
	/* true when the log was removed, due to a chain reorganization. false if its a valid log. */
	removed: boolean;
	/* Array of zero to four 32 Bytes DATA of indexed log arguments. */
	topics: string[];
	/* hash of the transaction */
	transactionHash: Hash;
  }